requirements.txt
colour==0.1.5
GitPython==3.1.26
numpy==1.21.3
pandas==1.4.2
compress_pickle==2.1.0
scikit_learn==1.0.2
setuptools==58.2.0
torch==1.11.0
tqdm==4.62.3
wandb==0.12.15
matplotlib==3.5.1
pillow==9.1.0
protobuf==3.20.1
